+++ src/game/tests.rs
@@ -977,6 +977,7 @@ mod spider_man;
 
 mod dragonfire_blade;
 mod lorwyn_eclipsed;
+mod ramos_dragon_engine;
 mod tarkir_dragonstorm;
 
 mod marvel_super_heroes;